Create README.md
It is played on a board of  9 squares, divided into 3 rows and 3 columns. Starting with an empty board, the players in alternating turns place a single piece in an empty square.
The starting player places X’s, the other player places O’s. The first player to line up 3 of their pieces in a line (horizontal, vertical, or diagonal), wins the game. If no more moves are possible the game ends in a draw